% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar_add_dimension.R
\name{add_dimension}
\alias{add_dimension}
\alias{add_dimension.multistar}
\title{Add a dimension table to a \code{multistar}}
\usage{
add_dimension(
  ms,
  dimension_name = NULL,
  dimension_table = NULL,
  dimension_key = NULL,
  fact_name = NULL,
  fact_key = NULL,
  key_as_data = FALSE
)

\method{add_dimension}{multistar}(
  ms,
  dimension_name = NULL,
  dimension_table = NULL,
  dimension_key = NULL,
  fact_name = NULL,
  fact_key = NULL,
  key_as_data = FALSE
)
}
\arguments{
\item{ms}{A \code{multistar} object.}

\item{dimension_name}{A string, name of dimension table.}

\item{dimension_table}{A \code{tibble}, dimension table.}

\item{dimension_key}{A string, name of the dimension primary key.}

\item{fact_name}{A string, name of fact table.}

\item{fact_key}{A string, name of the dimension foreign key.}

\item{key_as_data}{A boolean, define the primary key as an attribute of the
dimension accessible in queries?}
}
\value{
A \code{tibble}.
}
\description{
Add a dimension table to a \code{multistar}
}
\examples{
library(tidyr)


}
\seealso{


Other multistar functions: 
\code{\link{add_facts}()},
\code{\link{multistar}()},
\code{\link{relate_dimension}()}
}
\concept{multistar functions}
