% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar_add_facts.R
\name{add_facts}
\alias{add_facts}
\alias{add_facts.multistar}
\title{Add a fact table to a \code{multistar}}
\usage{
add_facts(ms, fact_name = NULL, fact_table = NULL)

\method{add_facts}{multistar}(ms, fact_name = NULL, fact_table = NULL)
}
\arguments{
\item{ms}{A \code{multistar} object.}

\item{fact_name}{A string, name of fact table.}

\item{fact_table}{A \code{tibble}, fact table.}
}
\value{
A \code{tibble}.
}
\description{
Add a fact table to a \code{multistar}
}
\examples{
library(tidyr)


}
\seealso{


Other multistar functions: 
\code{\link{add_dimension}()},
\code{\link{multistar}()},
\code{\link{relate_dimension}()}
}
\concept{multistar functions}
