% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{add_geodimension_additional_attributes}
\alias{add_geodimension_additional_attributes}
\title{Add geodimension additional attributes}
\usage{
add_geodimension_additional_attributes(dq)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Adds the geodimension attributes not explicitly selected in the query but
included in the definition.
}
\keyword{internal}
