% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{are_levels_related}
\alias{are_levels_related}
\title{Are levels related?}
\usage{
are_levels_related(level, upper_level)
}
\arguments{
\item{level}{A \code{geolevel}.}

\item{upper_level}{A \code{geolevel}.}
}
\value{
A boolean.
}
\description{
A level is always related to the \emph{Top} level.
}
\details{
A level can be added between any level and the \emph{Top} level, otherwise, lower
and upper levels must be related.
}
\keyword{internal}
