% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_check_referential_integrity.R
\name{check_referential_integrity}
\alias{check_referential_integrity}
\alias{check_referential_integrity.geolevel}
\title{Check the referential integrity of a level with respect to another}
\usage{
check_referential_integrity(
  level = NULL,
  upper_level = NULL,
  include_all_attributes = FALSE
)

\method{check_referential_integrity}{geolevel}(
  level = NULL,
  upper_level = NULL,
  include_all_attributes = FALSE
)
}
\arguments{
\item{level}{A \code{geolevel}, level to check.}

\item{upper_level}{A \code{geolevel}, reference level.}

\item{include_all_attributes}{A boolean, include all level attributes.}
}
\value{
A \code{tibble}.
}
\description{
Before relating one level to another in a dimension, we must check its
referential integrity: if all the instances of the lower level are going to
be able to relate to some instance of the higher level.
}
\details{
Returns a table with the instances that cannot be related. It can be selected
to include only the common attributes with the top level or all attributes on
the level.
}
\examples{

city <-
  geolevel(
    name = "city",
    layer = usa_cities,
    key = c("city", "state"),
    geometry = "point"
  )

state <-
  geolevel(
    name = "state",
    layer = usa_states,
    key = c("state"),
    geometry = "polygon"
  )

ft <- check_referential_integrity(city, state)

}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()}
}
\concept{geodimension functions}
