% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{complete_point_geometry}
\alias{complete_point_geometry}
\alias{complete_point_geometry.geolevel}
\title{complete point geometry}
\usage{
complete_point_geometry(gl)

\method{complete_point_geometry}{geolevel}(gl)
}
\arguments{
\item{gl}{A \code{geolevel} object.}
}
\value{
A \code{geolevel} object.
}
\description{
A level can have several associated geometries (point, polygon or line). Add
the geometry of the layer or replace an existing one of the indicated type.
}
\examples{
library(tidyr)


}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{complete_relation_by_geography}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry_instances}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()}
}
\concept{geodimension functions}
