% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_relate_levels.R
\name{complete_relation_by_geography}
\alias{complete_relation_by_geography}
\alias{complete_relation_by_geography.geodimension}
\title{complete relation by geography}
\usage{
complete_relation_by_geography(
  gd,
  lower_level_name = NULL,
  upper_level_name = NULL
)

\method{complete_relation_by_geography}{geodimension}(
  gd,
  lower_level_name = NULL,
  upper_level_name = NULL
)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{geodimension} object.
}
\description{
complete relation by geography
}
\examples{
library(tidyr)


}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{complete_point_geometry}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry_instances}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()}
}
\concept{geodimension functions}
