% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{default_attribute}
\alias{default_attribute}
\title{Default attribute}
\usage{
default_attribute(dq, dimension)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}

\item{dimension}{A string, name of the geographic dimension.}
}
\value{
A string, name of the attribute.
}
\description{
Get the attribute with the most instances, if there is more than one.
}
\keyword{internal}
