% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimensional_query_define_attribute_geom.R
\name{define_attribute_geom}
\alias{define_attribute_geom}
\alias{define_attribute_geom.geodimensional_query}
\title{Define geometric attribute}
\usage{
define_attribute_geom(
  gdq,
  dimension = NULL,
  attribute = NULL,
  sf = NULL,
  by = NULL
)

\method{define_attribute_geom}{geodimensional_query}(
  gdq,
  dimension = NULL,
  attribute = NULL,
  sf = NULL,
  by = NULL
)
}
\arguments{
\item{gdq}{A \code{geodimensional_query} object.}

\item{dimension}{A string, dimension name. By default it considers the first
geographic dimension.}

\item{attribute}{A string, attribute name.}

\item{sf}{A \code{sf} object}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} structure.}
}
\value{
A \code{geodimensional_query} object.
}
\description{
Define geometric attribute
}
\examples{
library(tidyr)


}
\seealso{

}
\concept{query functions}
