% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimensional_query_define_attribute_geom.R
\name{define_attribute_geom_from_attribute}
\alias{define_attribute_geom_from_attribute}
\title{Define one attribute from another}
\usage{
define_attribute_geom_from_attribute(
  gdq,
  dimension = NULL,
  attribute = NULL,
  from_attribute = NULL
)
}
\arguments{
\item{gdq}{A \code{geodimensional_query} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{from_attribute}{A string, attribute name.}
}
\value{
A \code{geodimensional_query} object.
}
\description{
Define one attribute from another.
}
\keyword{internal}
