% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimensional_query_define_attribute_geom.R
\name{define_attribute_geom_from_layer}
\alias{define_attribute_geom_from_layer}
\title{Define an attribute from a layer}
\usage{
define_attribute_geom_from_layer(
  gdq,
  dimension = NULL,
  attribute = NULL,
  layer = NULL,
  by = NULL
)
}
\arguments{
\item{gdq}{A \code{geodimensional_query} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{layer}{A \code{sf} object}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} structure.}
}
\value{
A \code{geodimensional_query} object.
}
\description{
Define an attribute from a layer.
}
\keyword{internal}
