% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomultistar_define_geoattribute.R
\name{define_geoattribute_from_layer}
\alias{define_geoattribute_from_layer}
\title{Define an attribute from a layer}
\usage{
define_geoattribute_from_layer(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL
)
}
\arguments{
\item{gms}{A \code{geomultistar} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{from_layer}{A \code{sf} object}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} structure.}
}
\value{
A \code{geomultistar} object.
}
\description{
Define an attribute from a layer.
}
\keyword{internal}
