% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{derelate_to_upper_level}
\alias{derelate_to_upper_level}
\title{De-relate level to upper level.}
\usage{
derelate_to_upper_level(level, upper_level_name)
}
\arguments{
\item{level}{A \code{geolevel}.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
level A \code{geolevel}.
}
\description{
De-relate level to upper level.
}
\keyword{internal}
