% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{geodimension}
\alias{geodimension}
\title{\code{geodimension} S3 class}
\usage{
geodimension(name = NULL, botton_level = NULL, top_level = NULL)
}
\arguments{
\item{name}{A string, name of the dimension.}

\item{botton_level}{A \code{geolevel}, level with the finest granularity.}

\item{top_level}{A \code{geolevel}, level \emph{All}, with a single instance.}
}
\value{
A \code{geodimension} object.
}
\description{
A \code{geodimension} object is created. Dimensions that contain geographic
information are indicated.
}
\examples{


}
\seealso{


Other geolevel functions: 
\code{\link{add_geometry}()},
\code{\link{geolevel}()}
}
\concept{geolevel functions}
