% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimensional_query.R
\name{geodimensional_query}
\alias{geodimensional_query}
\title{\code{geodimensional_query} S3 class}
\usage{
geodimensional_query(ms = NULL)
}
\arguments{
\item{ms}{A \code{multistar} object.}
}
\value{
A \code{geodimensional_query} object.
}
\description{
An empty \code{geodimensional_query} object is created where you can select fact
measures, dimension attributes and filter dimension rows.
}
\examples{
library(tidyr)
library(starschemar)

ms_mrs <- ct_mrs \%>\%
  constellation_as_multistar()

dq <- geodimensional_query(ms_mrs)

}
