% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{geolevel}
\alias{geolevel}
\title{\code{geolevel} S3 class}
\usage{
geolevel(
  name = NULL,
  layer = NULL,
  attributes = NULL,
  key = NULL,
  geometry = NULL,
  top_level = FALSE
)
}
\arguments{
\item{name}{A string, level name.}

\item{layer}{A \code{sf} object.}

\item{attributes}{A vector, selected attributes.}

\item{key}{A vector, attributes that compose the key.}

\item{geometry}{A string, type of geometry of the layer.}

\item{top_level}{A boolean, is the level Top of a dimension.}
}
\value{
A \code{geolevel} object.
}
\description{
A \code{geolevel} object is created. Dimensions that contain geographic
information are indicated.
}
\examples{


}
\seealso{


Other geolevel functions: 
\code{\link{add_geometry}()},
\code{\link{geodimension}()}
}
\concept{geolevel functions}
