% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{get_selected_measure_names}
\alias{get_selected_measure_names}
\title{get_selected_measure_names}
\usage{
get_selected_measure_names(dq, ft)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}

\item{ft}{A \code{tibble} object.}
}
\value{
A vector of names.
}
\description{
Get the names of the measures selected in the query.
}
\keyword{internal}
