% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_relate_levels.R
\name{get_unrelated_instances}
\alias{get_unrelated_instances}
\alias{get_unrelated_instances.geodimension}
\title{Relate levels in a dimension}
\usage{
get_unrelated_instances(gd, lower_level_name = NULL, upper_level_name = NULL)

\method{get_unrelated_instances}{geodimension}(gd, lower_level_name = NULL, upper_level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{tibble}.
}
\description{
Relate levels in a dimension
}
\examples{
library(tidyr)


}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry_instances}()},
\code{\link{relate_levels}()}
}
\concept{geodimension functions}
