% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R,
%   R/geomultistar_define_geoattribute.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_geolevel}
\alias{define_geoattribute_from_attribute}
\alias{define_geoattribute_from_layer}
\title{\code{geolevel} S3 class}
\usage{
new_geolevel(
  name = NULL,
  layer = NULL,
  attributes = NULL,
  key = NULL,
  geometry = NULL
)

define_geoattribute_from_attribute(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_attribute = NULL,
  additional_attributes = NULL
)

define_geoattribute_from_layer(
  gms,
  dimension = NULL,
  attribute = NULL,
  from_layer = NULL,
  by = NULL
)
}
\arguments{
\item{name}{A string, level name.}

\item{layer}{A \code{sf} object}

\item{attributes}{A vector, selected attributes.}

\item{key}{A vector, attributes that compose the key.}

\item{geometry}{A string, type of geometry of the layer.}

\item{gms}{A \code{geomultistar} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A string, attribute name.}

\item{from_attribute}{A string, attribute name.}

\item{additional_attributes}{A vector, attribute names.}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} structure.}
}
\value{
A \code{geolevel} object.

A \code{geomultistar} object.

A \code{geomultistar} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.

Define a geoattribute from another.

Define an attribute from a layer.
}
\keyword{internal}
