% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{is_top_level}
\alias{is_top_level}
\title{Is it \emph{Top} level of dimension?}
\usage{
is_top_level(level)
}
\arguments{
\item{level}{A \code{geolevel}.}
}
\value{
A boolean.
}
\description{
Is it \emph{Top} level of dimension?
}
\keyword{internal}
