% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrs_fact_cause}
\alias{mrs_fact_cause}
\title{Fact \emph{cause}}
\format{
A \code{tibble}.
}
\source{
\url{https://CRAN.R-project.org/package=starschemar}
}
\usage{
mrs_fact_cause
}
\description{
Fact \emph{cause} table of the Mortality Reporting System. Defined from \code{ms_mrs}.
Foreign keys have been renamed, only a \emph{when} dimension has been considered,
the type for the \emph{when} dimension has been changed.
}
\keyword{datasets}
