% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrs_when}
\alias{mrs_when}
\title{Dimension \emph{when}}
\format{
A \code{tibble}.
}
\source{
\url{https://CRAN.R-project.org/package=starschemar}
}
\usage{
mrs_when
}
\description{
\emph{When} dimension table of the Mortality Reporting System. Defined from
\code{ms_mrs}. The primary key has been renamed and its type has been changed. The
other attributes have also been renamed.
}
\keyword{datasets}
