% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrs_who}
\alias{mrs_who}
\title{Dimension \emph{who}}
\format{
A \code{tibble}.
}
\source{
\url{https://CRAN.R-project.org/package=starschemar}
}
\usage{
mrs_who
}
\description{
\emph{Who} dimension table of the Mortality Reporting System. Defined from
\code{ms_mrs}. The primary key has been renamed.
}
\keyword{datasets}
