% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ms_mrs_test}
\alias{ms_mrs_test}
\title{Multistar for Mortality Reporting System Test}
\format{
A \code{multistar} object.
}
\source{
\url{https://CRAN.R-project.org/package=starschemar}
}
\usage{
ms_mrs_test
}
\description{
Multistar for the Mortality Reporting System considering age and cause
classification data test.
}
\keyword{datasets}
