% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar.R
\name{multistar}
\alias{multistar}
\title{\code{multistar} S3 class}
\usage{
multistar()
}
\value{
A \code{multistar} object.
}
\description{
Given a name for the facts and the fact table, a \code{multistar} object is
created.
}
\examples{
library(tidyr)
library(starschemar)


}
\seealso{


Other multistar functions: 
\code{\link{add_dimension}()},
\code{\link{add_facts}()},
\code{\link{relate_dimension}()}
}
\concept{multistar functions}
