% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{name_with_nexus}
\alias{name_with_nexus}
\title{Name with nexus}
\usage{
name_with_nexus(name)
}
\arguments{
\item{name}{A string.}
}
\value{
A string.
}
\description{
Given a name, if it ends in "/" the nexus is the empty string, otherwise it
is "/". Add the nexus.
}
\keyword{internal}
