% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimensional_query.R
\name{new_geodimensional_query}
\alias{new_geodimensional_query}
\title{geodimensional_query S3 class}
\usage{
new_geodimensional_query(ms = NULL, geodimension = NULL)
}
\value{
A \code{geodimensional_query} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
