% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar.R
\name{new_multistar}
\alias{new_multistar}
\title{\code{multistar} S3 class}
\usage{
new_multistar(fact_name = NULL, fact_table = NULL)
}
\arguments{
\item{fact_name}{A string, name of fact table.}

\item{fact_table}{A \code{tibble}, fact table.}
}
\value{
A \code{multistar} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
