% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar.R
\name{new_multistar_empty}
\alias{new_multistar_empty}
\title{\code{multistar} S3 class}
\usage{
new_multistar_empty()
}
\value{
A \code{multistar} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
