% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{normalize_level}
\alias{normalize_level}
\alias{normalize_level.geolevel}
\title{Relate level to upper level.}
\usage{
normalize_level(level, upper_level)

\method{normalize_level}{geolevel}(level, upper_level)
}
\arguments{
\item{level}{A \code{geolevel}.}

\item{upper_level}{A \code{geolevel}.}
}
\value{
level A \code{geolevel}.
}
\description{
Relate level to upper level.
}
\keyword{internal}
