% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar_relate_dimension.R
\name{relate_dimension}
\alias{relate_dimension}
\alias{relate_dimension.multistar}
\title{Relate a dimension table to a fact table in a \code{multistar}}
\usage{
relate_dimension(ms, dimension_name = NULL, fact_name = NULL, fact_key = NULL)

\method{relate_dimension}{multistar}(ms, dimension_name = NULL, fact_name = NULL, fact_key = NULL)
}
\arguments{
\item{ms}{A \code{multistar} object.}

\item{dimension_name}{A string, name of dimension table.}

\item{fact_name}{A string, name of fact table.}

\item{fact_key}{A string, name of the dimension foreign key.}
}
\value{
A \code{tibble}.
}
\description{
Relate a dimension table to a fact table in a \code{multistar}
}
\examples{
library(tidyr)


}
\seealso{


Other multistar functions: 
\code{\link{add_dimension}()},
\code{\link{add_facts}()},
\code{\link{multistar}()}
}
\concept{multistar functions}
