% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{relate_to_top_level}
\alias{relate_to_top_level}
\alias{relate_to_top_level.geolevel}
\title{Relate level to \emph{Top} level}
\usage{
relate_to_top_level(level, top_level)

\method{relate_to_top_level}{geolevel}(level, top_level)
}
\arguments{
\item{level}{A \code{geolevel}.}

\item{top_level}{A \code{geolevel}, level \emph{All}, with a single instance.}
}
\value{
level A \code{geolevel}.
}
\description{
Relate level to \emph{Top} level
}
\keyword{internal}
