% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{relate_to_upper_level}
\alias{relate_to_upper_level}
\title{Relate level to upper level.}
\usage{
relate_to_upper_level(level, upper_level)
}
\arguments{
\item{level}{A \code{geolevel}.}

\item{upper_level}{A \code{geolevel}.}
}
\value{
level A \code{geolevel}.
}
\description{
Relate level to upper level.
}
\keyword{internal}
