% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{specify_geodimension}
\alias{specify_geodimension}
\title{Specify geodimension}
\usage{
specify_geodimension(dq, dimension, attribute)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}

\item{dimension}{A string, name of the geographic dimension.}

\item{attribute}{A string, name of the selected geographic attribute.}
}
\value{
A \code{dimensional_query} object.
}
\description{
Get the geodimension data.
}
\keyword{internal}
