% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_mrs_age_test}
\alias{st_mrs_age_test}
\title{Star Schema for Mortality Reporting System by Age Test}
\format{
A \code{star_schema} object.
}
\source{
\url{https://CRAN.R-project.org/package=starschemar}
}
\usage{
st_mrs_age_test
}
\description{
Star Schema for the Mortality Reporting System considering the age
classification data test.
}
\keyword{datasets}
