% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_acs.R
\name{tiger_acs}
\alias{tiger_acs}
\title{\code{tiger_acs} S3 class}
\usage{
tiger_acs(filepath = NULL, code = NULL)
}
\arguments{
\item{filepath}{A string, path to gbd file.}

\item{code}{A string.}
}
\value{
A \code{tiger_acs} object.
}
\description{
A \code{tiger_acs} object is created from a given
}
