% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_acs.R
\name{transform_code}
\alias{transform_code}
\title{Transform code}
\usage{
transform_code(mdr)
}
\arguments{
\item{mdr}{A \code{tibble} row.}
}
\value{
A \code{tibble} row.
}
\description{
Transform code
}
\keyword{internal}
