% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_acs.R
\name{transform_metadata}
\alias{transform_metadata}
\alias{transform_metadata.tiger_acs}
\title{Transform metadata}
\usage{
transform_metadata(ta)

\method{transform_metadata}{tiger_acs}(ta)
}
\arguments{
\item{ta}{A string.}
}
\value{
A \code{tiger_acs} object.
}
\description{
Transform metadata
}
\keyword{internal}
