% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_acs.R
\name{transform_values}
\alias{transform_values}
\title{Transform code}
\usage{
transform_values(mdr, values)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{values}{A vector of values}
}
\value{
A \code{tibble} row.
}
\description{
Transform code
}
\keyword{internal}
