% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_cities}
\alias{usa_cities}
\title{USA Cities, 2014}
\format{
A \code{sf}.
}
\source{
\url{https://earthworks.stanford.edu/catalog/stanford-bx729wr3020}
}
\usage{
usa_cities
}
\description{
From the original dataset, some fields have been selected and renamed, and
only includes the Mortality Reporting System cities.
}
\keyword{datasets}
