% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_counties}
\alias{usa_counties}
\title{USA Counties, 2018}
\format{
A \code{sf}.
}
\source{
\url{https://www2.census.gov/geo/tiger/GENZ2018/shp/cb_2018_us_county_20m.zip}
}
\usage{
usa_counties
}
\description{
From the original dataset, some fields have been selected and renamed.
}
\keyword{datasets}
