% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_regions}
\alias{usa_regions}
\title{USA Regions, 2018}
\format{
A \code{sf}.
}
\source{
\url{https://www2.census.gov/geo/tiger/GENZ2018/shp/cb_2018_us_region_20m.zip}
}
\usage{
usa_regions
}
\description{
From the original dataset, some fields have been selected and renamed.
}
\keyword{datasets}
