% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{widen_flat_table}
\alias{widen_flat_table}
\title{widen_flat_table}
\usage{
widen_flat_table(ft, pk, measures)
}
\arguments{
\item{ft}{A \code{tibble} object.}

\item{pk}{A vector of names.}

\item{measures}{A vector of names.}
}
\value{
A list of \code{tibble} objects.
}
\description{
Makes the pk fields the primary key of the table. The rest of the fields that
prevent them from being so, are extracted to another table. The original
table is widened by combining the extracted fields with the measures.
}
\keyword{internal}
