% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{add_geometry}
\alias{add_geometry}
\alias{add_geometry.geolevel}
\title{Add geometry to a level}
\usage{
add_geometry(gl, layer = NULL, geometry = NULL)

\method{add_geometry}{geolevel}(gl, layer = NULL, geometry = NULL)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{layer}{A \code{sf} object.}

\item{geometry}{A string, type of geometry of the layer.}
}
\value{
A \code{geolevel}.
}
\description{
A level can have several associated geometries (point, polygon or line). Add
the geometry of the layer or replace an existing one of the indicated type.
}
\examples{
library(tidyr)

city <-
  geolevel(
    name = "city",
    layer = usa_cities,
    attributes = NULL,
    key = c("city", "state"),
    geometry = "point"
  )

# geometry is no checked
city <- city \%>\%
  add_geometry(layer = usa_cities,
               geometry = "polygon")

}
\seealso{


Other geodimension functions: 
\code{\link{add_level}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()}
}
\concept{geodimension functions}
