% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{add_geometry}
\alias{add_geometry}
\alias{add_geometry.geolevel}
\title{Relate a dimension table to a fact table in a \code{geolevel}}
\usage{
add_geometry(gl, layer = NULL, geometry = NULL)

\method{add_geometry}{geolevel}(gl, layer = NULL, geometry = NULL)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{layer}{A \code{sf} object.}

\item{geometry}{A string, type of geometry of the layer.}
}
\value{
A \code{geolevel}.
}
\description{
Adding a dimension to a \code{geolevel} can only relate to a fact table. You can
then relate to other fact tables in the \code{geolevel} using this function. The
name of the fact table and its foreign key must be indicated. The referential
integrity of the instances of the facts is checked.
}
\examples{

}
\seealso{


Other geolevel functions: 
\code{\link{geodimension}()},
\code{\link{geolevel}()}
}
\concept{geolevel functions}
