% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_add_level.R
\name{add_level}
\alias{add_level}
\alias{add_level.geodimension}
\title{Relate a dimension table to a fact table in a \code{geodimension}}
\usage{
add_level(gd, level = NULL, lower_level_name = NULL, upper_level_name = NULL)

\method{add_level}{geodimension}(gd, level = NULL, lower_level_name = NULL, upper_level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level}{A \code{geolevel}, level to add to the dimension.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{geodimension}.
}
\description{
Adding a dimension to a \code{geodimension} can only relate to a fact table. You can
then relate to other fact tables in the \code{geodimension} using this function. The
name of the fact table and its foreign key must be indicated. The referential
integrity of the instances of the facts is checked.
}
\examples{

}
\seealso{

}
\concept{geodimension functions}
