% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_add_level.R
\name{add_level}
\alias{add_level}
\alias{add_level.geodimension}
\title{Add a level to a dimension}
\usage{
add_level(gd, level = NULL, lower_level_name = NULL, upper_level_name = NULL)

\method{add_level}{geodimension}(gd, level = NULL, lower_level_name = NULL, upper_level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level}{A \code{geolevel}, level to add to the dimension.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{geodimension}.
}
\description{
A new level must always lie between two levels of the dimension. Given the
level and the names of the two levels between which it is inserted, it adds
it to the dimension.
}
\details{
If any of them is not indicated, by default it is considered the lowest level
and the highest level of the dimension.
}
\examples{
library(tidyr)

city <-
  geolevel(
    name = "city",
    layer = usa_cities,
    key = c("city", "state"),
    geometry = "point"
  )

all <-
  geolevel(
    layer = usa_nation,
    geometry = "polygon",
    top_level = TRUE
  )

usa_city <-
  geodimension(name = "usa_city",
               botton_level = city,
               top_level = all)

state <-
  geolevel(
    name = "state",
    layer = usa_states,
    key = c("state"),
    geometry = "polygon"
  )

usa_city <- usa_city \%>\%
  add_level(level = state)

}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()}
}
\concept{geodimension functions}
