% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimensional_query_define_attribute_geom.R
\name{define_attribute_geom}
\alias{define_attribute_geom}
\alias{define_attribute_geom.geodimensional_query}
\title{Define geometric attribute}
\usage{
define_attribute_geom(
  gdq,
  attribute = NULL,
  layer = NULL,
  by = NULL,
  from_attribute = NULL
)

\method{define_attribute_geom}{geodimensional_query}(
  gdq,
  attribute = NULL,
  layer = NULL,
  by = NULL,
  from_attribute = NULL
)
}
\arguments{
\item{gdq}{A \code{geodimensional_query} object.}

\item{attribute}{A string, attribute name.}

\item{layer}{A \code{sf} object}

\item{by}{a vector of correspondence of attributes of the dimension with the
\code{sf} structure.}

\item{from_attribute}{A string, attribute name.}
}
\value{
A \code{geodimensional_query} object.
}
\description{
Define geometric attribute
}
\examples{
library(tidyr)


}
\seealso{

}
\concept{query functions}
