% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{geodimension}
\alias{geodimension}
\title{\code{geodimension} S3 class}
\usage{
geodimension(name = NULL, botton_level = NULL, top_level = NULL)
}
\arguments{
\item{name}{A string, name of the dimension.}

\item{botton_level}{A \code{geolevel}, level with the finest granularity.}

\item{top_level}{A \code{geolevel}, level \emph{All}, with a single instance.}
}
\value{
A \code{geodimension} object.
}
\description{
A \code{geodimension} object is created. In addition to the name of the dimension,
the lowest and highest level of the dimension are indicated.
}
\examples{

city <-
  geolevel(
    name = "city",
    layer = usa_cities,
    key = c("city", "state"),
    geometry = "point"
  )

all <-
  geolevel(
    layer = usa_nation,
    geometry = "polygon",
    top_level = TRUE
  )

usa_city <-
  geodimension(name = "usa_city",
               botton_level = city,
               top_level = all)

}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{geolevel}()}
}
\concept{geodimension functions}
