% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{geodimension}
\alias{geodimension}
\title{\code{geodimension} S3 class}
\usage{
geodimension(name = NULL, level = NULL)
}
\arguments{
\item{name}{A string, name of the dimension.}

\item{level}{A \code{geolevel}.}
}
\value{
A \code{geodimension} object.
}
\description{
A \code{geodimension} object is created. In addition to the name of the dimension,
the lowest and highest level of the dimension are indicated.
}
\examples{

city <-
  geolevel(
    name = "city",
    layer = usa_cities,
    key = c("city", "state"),
    geometry = "point"
  )

usa_city <-
  geodimension(name = "usa_city",
               level = city)

}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()}
}
\concept{geodimension functions}
