% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{geolevel}
\alias{geolevel}
\title{\code{geolevel} S3 class}
\usage{
geolevel(
  name = NULL,
  layer = NULL,
  attributes = NULL,
  key = NULL,
  geometry = NULL,
  top_level = FALSE
)
}
\arguments{
\item{name}{A string, level name.}

\item{layer}{A \code{sf} object.}

\item{attributes}{A vector, selected attributes.}

\item{key}{A vector, attributes that compose the key.}

\item{geometry}{A string, type of geometry of the layer.}

\item{top_level}{A boolean, is the level \emph{Top} of a dimension.}
}
\value{
A \code{geolevel} object.
}
\description{
A \code{geolevel} object is created from a given geographic layer. The attributes
of the layer to be included in the level are indicated, and the subset of
these that make up the natural key. If no attribute is indicated, all are
considered.
}
\details{
A level can have several associated geometries (point, polygon or line), the
geometry of the layer is indicated.

The \emph{Top} level of any dimension is a special case since it has only one
object. This situation is indicated by a boolean parameter.
}
\examples{

city <-
  geolevel(
    name = "city",
    layer = usa_cities,
    key = c("city", "state"),
    geometry = "point"
  )

all <-
  geolevel(
    layer = usa_nation,
    geometry = "polygon",
    top_level = TRUE
  )

}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{geodimension}()}
}
\concept{geodimension functions}
