% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistar.R
\name{multistar}
\alias{multistar}
\title{\code{multistar} S3 class}
\usage{
multistar(fact_name = NULL, fact_table = NULL)
}
\arguments{
\item{fact_name}{A string, name of fact table.}

\item{fact_table}{A \code{tibble}, fact table.}
}
\value{
A \code{multistar} object.
}
\description{
Given a name for the facts and the fact table, a \code{multistar} object is
created.
}
\examples{
library(tidyr)
library(starschemar)


}
\seealso{


Other geo functions: 
\code{\link{define_geoattribute}()},
\code{\link{geomultistar}()},
\code{\link{get_empty_geoinstances}()},
\code{\link{run_geoquery}()}
}
\concept{geo functions}
