% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{new_geodimension}
\alias{new_geodimension}
\title{\code{geodimension} S3 class}
\usage{
new_geodimension(name = NULL, botton_level = NULL, top_level = NULL)
}
\arguments{
\item{name}{A string, name of the dimension.}

\item{botton_level}{A \code{geolevel}, level with the finest granularity.}

\item{top_level}{A \code{geolevel}, level \emph{All}, with a single instance.}
}
\value{
A \code{geodimension} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
