% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_run_geoquery.R
\name{specify_geodimension}
\alias{specify_geodimension}
\title{Default attribute}
\usage{
specify_geodimension(dq, dimension, attribute)
}
\arguments{
\item{dq}{A \code{dimensional_query} object.}

\item{dimension}{A string, name of the geographic dimension.}

\item{attribute}{A string, name of the selected geographic attribute.}
}
\value{
A string, name of the attribute.
}
\description{
Get the attribute with the most instances, if there is more than one.
}
\keyword{internal}
